backend/requirements.txt
flask
pillow
opencv-python
numpy
stegano
pydub
uuid

